Moving remote-connection back to root yml
Moving the remote-connection instructions into a role did not work.
I think this is because roles do not expect hosts files to be mentioned
in them. So, moving this back to the root level.

Am keeping a role around, doing a ping as a test to see if I can specify
roles in the rest of the root file.